changed class to operations functions
