--- synapseclient/models/protocols/download_list_protocol.py
@@ -1,240 +1,0 @@
-"""Protocol for the specific methods of this class that have synchronous counterparts
-generated at runtime."""
-
-from typing import TYPE_CHECKING, Optional, Protocol
-
-from synapseclient.core.async_utils import async_to_sync
-
-if TYPE_CHECKING:
-    from synapseclient import Synapse
-    from synapseclient.models.download_list import DownloadListItem
-    from synapseclient.models.table_components import CsvTableDescriptor
-
-
-@async_to_sync
-class DownloadListSynchronousProtocol(Protocol):
-    """
-    The protocol for methods that are asynchronous but also
-    have a synchronous counterpart that may also be called.
-    """
-
-    def download_files(
-        self,
-        download_location: Optional[str] = None,
-        *,
-        parallel: bool = False,
-        max_concurrent: int = 10,
-        synapse_client: Optional["Synapse"] = None,
-    ) -> str:
-        """Download all files in the Synapse download list (cart) to a local directory.
-
-        Files are downloaded individually. The cart is not packaged into a zip because
-        download lists can exceed 100 GB. Only successfully downloaded files are removed
-        from the cart after the full pass completes, so interrupted runs are safely
-        resumable.
-
-        Arguments:
-            download_location: Directory to download files to. Defaults to the
-                current working directory.
-            parallel: If True, files are downloaded concurrently up to
-                max_concurrent at a time using asyncio.gather. If False
-                (default), files are downloaded sequentially.
-            max_concurrent: Maximum number of files to download concurrently when
-                parallel=True. Defaults to 10. Has no effect when
-                parallel=False.
-            synapse_client: If not passed in and caching was not disabled by
-                Synapse.allow_client_caching(False) this will use the last created
-                instance from the Synapse class constructor.
-
-        Returns:
-            Path to the result manifest CSV, which contains all original manifest
-            columns plus path (local file path) and error (error message or
-            empty string) columns.
-
-        Example: Download all files in the cart
-            &nbsp;
-            Download all files in the user's download list to a local directory.
-            ```python
-            from synapseclient.models import DownloadList
-            from synapseclient import Synapse
-
-            syn = Synapse()
-            syn.login()
-
-            manifest_path = DownloadList.download_files(download_location="./downloads")
-            ```
-
-        Example: Download files in parallel
-            &nbsp;
-            Use parallel=True with max_concurrent for faster downloads on large carts.
-            ```python
-            from synapseclient.models import DownloadList
-            from synapseclient import Synapse
-
-            syn = Synapse()
-            syn.login()
-
-            manifest_path = DownloadList.download_files(
-                download_location="./downloads",
-                parallel=True,
-                max_concurrent=5,
-            )
-            ```
-        """
-        return ""
-
-    def get_manifest(
-        self,
-        *,
-        csv_table_descriptor: Optional["CsvTableDescriptor"] = None,
-        destination: str = ".",
-        synapse_client: Optional["Synapse"] = None,
-    ) -> str:
-        """Generate and download the download list manifest CSV without downloading files.
-
-        Arguments:
-            csv_table_descriptor: Optional CsvTableDescriptor controlling the
-                format of the generated CSV (separator, quote character, escape
-                character, line ending, and whether the first line is a header).
-                When omitted the Synapse defaults are used.
-            destination: Directory to download the manifest CSV to. Defaults to
-                the current working directory.
-            synapse_client: If not passed in and caching was not disabled by
-                Synapse.allow_client_caching(False) this will use the last created
-                instance from the Synapse class constructor.
-
-        Returns:
-            Local path to the downloaded manifest CSV.
-
-        Example: Get the download list manifest
-            &nbsp;
-            Inspect the cart contents before downloading.
-            ```python
-            from synapseclient.models import DownloadList
-            from synapseclient import Synapse
-
-            syn = Synapse()
-            syn.login()
-
-            manifest_path = DownloadList.get_manifest()
-            ```
-
-        Example: Get a manifest with custom CSV formatting
-            &nbsp;
-            Generate a manifest with custom separator, quote, escape, line
-            ending, and header settings.
-            ```python
-            from synapseclient.models import DownloadList
-            from synapseclient.models.table_components import CsvTableDescriptor
-            from synapseclient import Synapse
-
-            syn = Synapse()
-            syn.login()
-
-            descriptor = CsvTableDescriptor(
-                separator="\\t",
-                quote_character="'",
-                escape_character="\\\\",
-                line_end="\\n",
-                is_first_line_header=False,
-            )
-            manifest_path = DownloadList.get_manifest(csv_table_descriptor=descriptor)
-            ```
-        """
-        return ""
-
-    def add_files(
-        self,
-        files: list["DownloadListItem"],
-        *,
-        synapse_client: Optional["Synapse"] = None,
-    ) -> int:
-        """Add specific file versions to the Synapse download list.
-
-        Arguments:
-            files: List of DownloadListItem objects identifying file versions to add.
-            synapse_client: If not passed in and caching was not disabled by
-                Synapse.allow_client_caching(False) this will use the last created
-                instance from the Synapse class constructor.
-
-        Returns:
-            The number of files added.
-
-        Example: Add files to the download list
-            &nbsp;
-            Add specific file versions to the cart.
-            ```python
-            from synapseclient.models import DownloadList, DownloadListItem
-            from synapseclient import Synapse
-
-            syn = Synapse()
-            syn.login()
-
-            count = DownloadList.add_files([
-                DownloadListItem(file_entity_id="syn123", version_number=1),
-                DownloadListItem(file_entity_id="syn456", version_number=2),
-            ])
-            ```
-        """
-        return 0
-
-    def remove_files(
-        self,
-        files: list["DownloadListItem"],
-        *,
-        synapse_client: Optional["Synapse"] = None,
-    ) -> int:
-        """Remove specific file versions from the Synapse download list.
-
-        Arguments:
-            files: List of DownloadListItem objects identifying file versions to remove.
-            synapse_client: If not passed in and caching was not disabled by
-                Synapse.allow_client_caching(False) this will use the last created
-                instance from the Synapse class constructor.
-
-        Returns:
-            The number of files removed.
-
-        Example: Remove files from the download list
-            &nbsp;
-            Remove specific file versions from the cart.
-            ```python
-            from synapseclient.models import DownloadList, DownloadListItem
-            from synapseclient import Synapse
-
-            syn = Synapse()
-            syn.login()
-
-            count = DownloadList.remove_files([
-                DownloadListItem(file_entity_id="syn123", version_number=1),
-            ])
-            ```
-        """
-        return 0
-
-    def clear(
-        self,
-        *,
-        synapse_client: Optional["Synapse"] = None,
-    ) -> None:
-        """Clear all files from the Synapse download list (cart).
-
-        Arguments:
-            synapse_client: If not passed in and caching was not disabled by
-                Synapse.allow_client_caching(False) this will use the last created
-                instance from the Synapse class constructor.
-
-        Example: Clear the download list
-            &nbsp;
-            Remove all files from the cart.
-            ```python
-            from synapseclient.models import DownloadList
-            from synapseclient import Synapse
-
-            syn = Synapse()
-            syn.login()
-
-            DownloadList.clear()
-            ```
-        """
-        return
